requirements.txt
django
channels
daphne
django-environ
channels_redis
djangorestframework
